Use Arrays.equals
This is a behavior change. [].equals is equivalent to == (reference comparison), which does not seem to be what is intended by the String comparison on the framework method name. Arrays.equals is now used to compare the Class items in the arrays for equality.

Note that some users cannot currently compile from source as .equals on arrays is configured to be a compile-time error.