Pass the most specific Description to rules.
This is a behavior change. Prior to this change, rules for
parameterised tests received identical descriptions which
differed from the decription presented to the external
test runner. After this change, rules will receive the
more unique description that is presented to the external
test runner.

For example:

External test runner receives:
- rulesShouldReceiveCorrectDescription
  - [0] 0 (rulesShouldReceiveCorrectDescription)
  - [1] 1 (rulesShouldReceiveCorrectDescription)

Rules receive:

**Before**
- rulesShouldReceiveCorrectDescription
- rulesShouldReceiveCorrectDescription

**After**
- [0] 0 (rulesShouldReceiveCorrectDescription
- [1] 1 (rulesShouldReceiveCorrectDescription